Extracts renaming/creation of alchemy_roles column to own migration.
When upgrading an app to Alchemy 3.0 it comes with an alchemy_users table. So the first migration of this gem (creation of alchemy_users) will not run. But the existing roles column of the app has to be updated to alchemy_roles anyway. That's the reason why we moved that column changing part to a separate migration and check if we need to rename or add the alchemy_roles column.