Require password reset for users signing in the first time
By storing the `password_changed_at` time and by checking the `sign_in_count` we require users that sign in the first time to reset their password.

This only effects users that get created by other users either through the console, a seeder or by using the admin users form. Users signing up through the sign up form are not affected, because the sign up counts as first sign in.

If you want to opt out of this you simply can set `password_changed_at` to a date in the past in your seeder.